+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Respect favorited value in EmbeddedTweetView schibsted/tweet_ui#52
+
 ## [2.2.0+2] - 17.07.2020
 ### Changed
 - Fix formating